requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
idna==2.10
image==1.5.33
Pillow==8.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4




